Do not create matrixes, if they are already created

+++ matrix.py
@@ -2,13 +2,17 @@ import numpy as np
 import gc
 import time
 import sys
+import os.path
+# from memory_profiler import profile
+
 start_time = time.time()
 
+# @profile
 def calculateMatrix(Ap=None, Bp=None, Cp=None):
   print("started at %s" % start_time)
   # Sizes of matrices according to exercise definition
   if (Ap is None and Bp is None and Cp is None):
-    matrix_size_one = 10**5
+    matrix_size_one = 10**6
     matrix_size_two = 10**3
 
     # Matrice shapes
@@ -16,23 +20,25 @@ def calculateMatrix(Ap=None, Bp=None, Cp=None):
     B_shape = (matrix_size_two, matrix_size_one)
     C_shape = (matrix_size_one, 1)
 
-    # Create Matrix A, B, C into memory, (0,1), i.e. no 0 or 1 accepted
-    # np.random.rand can't be used, as its [0,1)
-    A = np.random.uniform(low=0.0001, high=0.9999, size=A_shape)
-    B = np.random.uniform(low=0.0001, high=0.9999, size=B_shape)
-    C = np.random.uniform(low=0.0001, high=0.9999, size=C_shape)
-    print("size of A in mem", sys.getsizeof(A))
+    if not (os.path.exists('data/matrix_a.npy') and os.path.exists('data/matrix_b.npy') and os.path.exists('data/matrix_c.npy')):
+      # Create Matrix A, B, C into memory, (0,1), i.e. no 0 or 1 accepted
+      # np.random.rand can't be used, as its [0,1)
+      A = np.random.uniform(low=0.0001, high=0.9999, size=A_shape)
+      B = np.random.uniform(low=0.0001, high=0.9999, size=B_shape)
+      C = np.random.uniform(low=0.0001, high=0.9999, size=C_shape)
+      print("size of A in mem", sys.getsizeof(A))
 
-    # Save matrices as data on hard drive, Matrix A and B will allocate 10^6*10^3*8 bytes of space, which is 8 Gigabytes
-    # These need to be on hard disk
-    np.save('data/matrix_a.npy', A)
-    np.save('data/matrix_b.npy', B)
-    np.save('data/matrix_c.npy', C)
-    # Call explicitely Python garbage collect, to remove A, B, C from memory
-    del A, B, C
-    gc.collect()
+      # Save matrices as data on hard drive, Matrix A and B will allocate 10^6*10^3*8 bytes of space, which is 8 Gigabytes
+      # These need to be on hard disk
+      np.save('data/matrix_a.npy', A)
+      np.save('data/matrix_b.npy', B)
+      np.save('data/matrix_c.npy', C)
 
-    # Fetch the data from hard disk to memory when needed, out-of-core algorithm
+      # Call explicitely Python garbage collect, to remove A, B, C from memory
+      del A, B, C
+      gc.collect()
+
+    # Fetch the data from hard disk to memory when needed, out-of-core algorithm idea
     A_mem = np.load('data/matrix_a.npy', mmap_mode='r+')
     B_mem = np.load('data/matrix_b.npy', mmap_mode='r+')
     C_mem = np.load('data/matrix_c.npy', mmap_mode='r+')
@@ -42,7 +48,7 @@ def calculateMatrix(Ap=None, Bp=None, Cp=None):
     B_mem = Bp
     C_mem = Cp
   # Set the batch size
-  batch_size = 1000
+  batch_size = 2000
 
   # Number of batches of matrix A and B
   A_row_batches = A_mem.shape[0] // batch_size
@@ -54,7 +60,10 @@ def calculateMatrix(Ap=None, Bp=None, Cp=None):
   for i in range(A_row_batches):
     # Collect all rows from batches
     temp_rows = [[] for _ in range(batch_size)]
+    cycle_start = time.time()
+    print("i", i)
     for j in range(B_column_batches):
+      # print("j", j)
       # Select sub matrix rows in batches from 2D A_mem, and all columns
       block_a = A_mem[i*batch_size:(i+1)*batch_size,:]
       # Select sub matrix columns in batches from 2D B_mem, and all rows
@@ -69,10 +78,13 @@ def calculateMatrix(Ap=None, Bp=None, Cp=None):
       gc.collect()
 
     # For each batch temporary_row, we want to multiply with C
-    print(i)
     for row in temp_rows:
       with_c = np.dot(row, C_mem)
       result = np.append(result, with_c)
+      del with_c
+    del temp_rows
+    print("--- %s seconds ---" % (time.time() - cycle_start))
+    gc.collect()
 
   # Make the result a 2D array and reshape it to be dimension x,1
   return_value = np.array([result]).reshape(-1, 1)
